% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMissing-package.R
\docType{package}
\name{plotMissing-package}
\alias{plotMissing-package}
\alias{_PACKAGE}
\title{plotMissing: Plot Missing And Observed Values}
\description{
Provides functions to easily generate customizable visualizations of the number of missing (`NA`) and observed (non-`NA`) values for each selected column in a tibble object. The functions in this package are designed to interact with the `tidyverse` set of packages.
}
\author{
\strong{Maintainer}: Joey Hotz \email{johotz37@gmail.com}

}
\keyword{internal}
