% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMissing.R
\name{plotMissing}
\alias{plotMissing}
\title{Plot counts of missing and recorded values in each selected column}
\usage{
plotMissing(
  .data,
  ...,
  cols = tidyselect::everything(),
  count_names = c("Missing", "Recorded"),
  bar_colours = c("red", "darkgreen")
)
}
\arguments{
\item{.data}{A tibble in R which contains data. This parameter is named \code{.data} because it is a generic name for a dataset, and this name matches the convention of other tidyverse functions. The \code{.} allows this function to be called using a pipe, as opposed to requiring the dataset to be explicitly given in the function call.}

\item{...}{A collection of additional graphical parameters which will be passed to \code{ggplot2}. This parameter gives us extra flexibility in creating the plots by dding desired graphical features after the creation of the "basic" plot.}

\item{cols}{<\href{https://tidyr.tidyverse.org/reference/tidyr_tidy_select.html}{\code{tidy-select}}> The columns in \code{.data} which will be shown in the plot. This parameter is named \code{cols} because it is a selection of columns, and this name matches the convention of other tidyverse functions.}

\item{count_names}{A vector of two strings, which will be the names given to the counts of missing and recorded values, respectively. This parameter is named \code{count_names} because the two values in this vector are the names corresponding to the counts of missing and recorded values.}

\item{bar_colours}{A vector of two colour codes, which are used for depicting the missing and recorded values, respectively. For more information on valid colour codes, see \code{colors()}. This parameter is named \code{bar_colours} because that is the proper Canadian spelling, and because the two values in this vector are the colours in the plot which correspond to the missing missing and recorded values.}
}
\value{
A stacked column plot which displays the number of missing and recorded observations per column
in \code{cols}. This plot has a legend with the group names given in \code{count_names}, where the colours
for each group are provided in \code{bar_colours}, and the plot also includes any additional
graphical parameters which were provided using the \code{...} argument.
}
\description{
This function returns a stacked column plot which depicts the number of missing (\code{NA})
and recorded (non-\code{NA}) observations in each specified column of a tibble.
}
\examples{
# This function does not require every single tidyverse library
# but it is convenient to load them all at once
library(tidyverse)
# load in a built-in dataset in R and convert it to a tibble
airquality <- tibble::as_tibble(airquality)
# displays a plot with the number of missing/recorded observations in each column
plotMissing(airquality)
# subsetting columns, changing the group names, and changing the bar colours
plotMissing(airquality,
            cols = c(Ozone, Solar.R),
            count_names = c("No", "Yes"),
            bar_colours = c("purple", "orange"))
# adding additional graphical parameters (axis labels, theme, and centering the title)
plotMissing(airquality,
            labs(x = "Column Name", y = "Number of Rows",
                 title = "Count of Missing Values in Selected Columns"),
            theme_bw(),
            theme(plot.title = element_text(hjust = 0.5)),
            cols = c(Ozone, Solar.R),
            count_names = c("No", "Yes"),
            bar_colours = c("purple", "red"))

}
