% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMissing-package.R
\docType{package}
\name{plotMissing-package}
\alias{plotMissing-package}
\alias{_PACKAGE}
\title{plotMissing: Plot Missing And Observed Values}
\description{
Generates a stacked column plot depicting the number of missing (`NA`) and observed (non-`NA`) values for each selected column in a tibble. The function provides built-in customization options including changing the colours of the bars, the group names in the legend, and selecting a subset of the columns to plot.
}
\author{
\strong{Maintainer}: Joey Hotz \email{johotz37@gmail.com}

}
\keyword{internal}
